+++ ui/mods/keyboard_viewer/keyboard_viewer.js
@@ -36,8 +36,8 @@ define([
 
       if (board == '') board = 'normal:'
       if (item.options.set == 'terrain editor') board = 'terrain:' + board
-      if (item.options.display_sub_group == '!LOC(settings:free_movement.message):free movement') board = 'freecam:' + board
-      if (item.options.display_sub_group == '!LOC(settings:build_items.message):build items') board = 'build:' + board
+      if (item.options.display_sub_group == '!LOC:free movement') board = 'freecam:' + board
+      if (item.options.display_sub_group == '!LOC:build items') board = 'build:' + board
 
       boards[board] = boards[board] || {}
       boards[board][key] = item